pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
"discord.py" = "^2.4.0"
psutil = "^6.0.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pytz = "^2024.1"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
